Fixed SystemNetworkContextManagerBrowsertest.CheckForBraveServiceKey test
